An htmlwidget interface to the d3po javascript chart library

Apache licensed alternative to 'Highcharter' which provides functions for both fast and beautiful interactive visualization for 'Markdown' and 'Shiny'.

This function provides 'd3po' methods from R console

Usage

d3po(data = NULL, ..., width = NULL, height = NULL, elementId = NULL)

Arguments

data

d3po need explicit specified data objects formatted as JSON, and this parameter passed it from R.

...

Aesthetics to pass, see daes()

width

Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

height

Same as width parameter.

elementId

Dummy string parameter. Useful when you have two or more charts on the same page.

Value

Creates a basic 'htmlwidget' object for simple visualization

See also

Author

Maintainer: Mauricio Vargas Sepulveda m.vargas.sepulveda@gmail.com (ORCID) [copyright holder]

Authors:

John Coene

Other contributors:

Ariel Alvarado [contributor]

Sylvain Lesage [contributor]

Curran Kelleher [contributor]

Fernando Becerra [contributor]

Natural Earth [data contributor]

R Consortium (Funded for the 2016-2017 ISC grants cycle) [funder]