Home 🏠

Functions daes po_box po_treemap po_pie po_donut po_area po_bar po_line po_scatter po_tooltip po_format po_labels po_font po_theme po_download po_network po_geomap d3po d3po_output render_d3po d3po_proxy

Vignettes 'd3po' worked examples

News Changelog

d3po {d3po}	R Documentation

An htmlwidget interface to the d3po javascript chart library

Description

Apache licensed alternative to 'Highcharter' which provides functions for both fast and beautiful interactive visualization for 'Markdown' and 'Shiny'.

Provides 'd3po' methods from R console

Usage

d3po(data = NULL, ..., width = NULL, height = NULL, element_id = NULL)

Arguments

data 	 d3po need explicit specified data objects formatted as JSON, and this parameter passed it from R.

... 	 Aesthetics to pass, see daes()

width 	 Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

height 	 Same as width parameter.

element_id 	 Dummy string parameter. Useful when you have two or more charts on the same page.

Value

Creates a basic 'htmlwidget' object for simple visualization

Author(s)

Maintainer: Mauricio Vargas Sepulveda m.vargas.sepulveda@gmail.com (ORCID) [copyright holder]

Authors:

John Coene

Other contributors:

Ariel Alvarado [contributor]

Sylvain Lesage [contributor]

Curran Kelleher [contributor]

Fernando Becerra [contributor]

Natural Earth [data contributor]

R Consortium (Funded for the 2016-2017 ISC grants cycle) [funder]

See Also

Useful links:

https://pacha.dev/d3po/

Report bugs at https://github.com/pachadotdev/d3po/issues