Shiny bindings for 'd3po'

Output and render functions for using d3po within Shiny applications and interactive Rmd documents.

d3po_output(output_id, width = "100%", height = "400px")

render_d3po(expr, env = parent.frame(), quoted = FALSE)

d3po_proxy(id, session = shiny::getDefaultReactiveDomain())

Arguments

output_id
output variable to read from

width, height
Must be a valid CSS unit (like '100%', '400px', 'auto') or a number, which will be coerced to a string and have 'px' appended.

expr
An expression that generates a d3po object

env
The environment in which to evaluate expr.

quoted
Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.

id
Id of plot to create a proxy of.

session
A valid shiny session.

Value

Creates a basic 'htmlwidget' object for 'Shiny' and interactive documents