Geomap

Plot a geomap

po_geomap(d3po, ..., data = NULL, map = NULL, inherit_daes = TRUE)

Arguments

d3po

Either the output of d3po() or d3po_proxy().

...

Aesthetics, see daes().

data

Any dataset to use for plot, overrides data passed to d3po().

map

map to use (i.e., any valid list or topojson file such as maps$south_america$chile or jsonlite::fromJSON("chile.topojson", simplifyVector = F))

inherit_daes

Whether to inherit aesthetics previous specified.

Value

an 'htmlwidgets' object with the desired interactive plot

Examples

# Get region IDs and names from the Chile map
dout <- map_ids(d3po::maps$south_america$chile)

# Only Magallanes has Pokemon (Mewtwo lives there!)
dout$pokemon_count <- ifelse(dout$id == "MA", 1L, 0L)
dout$color <- ifelse(dout$id == "MA", "#F85888", "#e0e0e0")

d3po(dout) %>%
  po_geomap(
    daes(
      group = id, color = color, size = pokemon_count,
      tooltip = name
    ),
    map = d3po::maps$south_america$chile
  ) %>%
  po_title("Pokemon Distribution in Chile")